requirements.txt
asgiref==3.3.1
bcrypt==3.2.0
cffi==1.14.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.6
django-import-export==2.5.0
django-model-utils==4.1.1
djangorestframework==3.12.2
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.6
pkg-resources==0.0.0
pycparser==2.20
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
xlrd==2.0.1
xlwt==1.3.0
